+++ tests/objectcsv/test_create_csv.py
@@ -136,8 +136,34 @@ def test_create_csv(datadir, test_config):
         data = list(reader)
         assert len(data) == len(["DC.xml", "SOURCE.xml"])
         assert data[0]["dsid"] == "DC.xml"
+        assert data[0]["dspath"] == "DC.xml"
         assert data[1]["dsid"] == "SOURCE.xml"
+        assert data[1]["dspath"] == "SOURCE.xml"
         assert data[0]["mimetype"] == "application/xml"
+
+
+def test_create_csv_with_subdirectories(datadir, test_config):
+    """Test the create_csv function on an object directory with subdirectories."""
+    # we move the test jpeg into a image folder befor calling create_csv
+    object_dir = datadir / "objects" / "obj1"
+    xml_dir = object_dir / "xml"
+    xml_dir.mkdir()
+    xml_file = object_dir / "SOURCE.xml"
+    xml_file.rename(xml_dir / "SOURCE.xml")  # move the file to the object directory
+    assert (xml_dir / "SOURCE.xml").exists()
+
+    create_csv(object_dir, test_config)
+
+    # check contents of the newly created datastreams.csv file
+    ds_csv = object_dir / "datastreams.csv"
+    with open(ds_csv, encoding="utf-8", newline="") as f:
+        reader = csv.DictReader(f)
+        data = list(reader)
+        assert len(data) == len(["DC.xml", "xml/SOURCE.xml"])
+        assert data[0]["dsid"] == "DC.xml"
+        assert data[0]["dspath"] == "DC.xml"
+        assert data[1]["dsid"] == "SOURCE.xml"
+        assert data[1]["dspath"] == "xml/SOURCE.xml"
 
 
 def test_create_csv_force_overwrite(datadir, test_config):